make the program modular
