+++ program/src/instruction/initialize.rs
@@ -22,11 +22,10 @@ fn initialize(accounts: &[AccountInfo], authorized: Authorized, lockup: Lockup) 
     // `get_stake_state()` is called unconditionally, which checks owner
     // do_initialize(stake_account_info, authorized, lockup, rent)?;
 
+    let rent = &Rent::from_account_info(rent_info)?;
 
-        let rent = &Rent::from_account_info(rent_info)?;
+    // `get_stake_state()` is called unconditionally, which checks owner
+    // do_initialize(stake_account_info, authorized, lockup, rent)?;
 
-        // `get_stake_state()` is called unconditionally, which checks owner
-        do_initialize(stake_account_info, authorized, lockup, rent)?;
-
-        Ok(())
-    }
+    Ok(())
+}
