+++ program/src/instruction/split.rs
@@ -1,7 +1,4 @@
-use crate::{
-    helpers::*,
-    state::{StakeHistorySysvar, StakeStateV2},
-};
+use crate::{helpers::*, state::stake_state_v2::StakeStateV2, state::StakeHistorySysvar};
 use pinocchio::{
     account_info::AccountInfo,
     program_error::ProgramError,
