+++ program/src/lib.rs
@@ -6,14 +6,14 @@ mod entrypoint;
 extern crate std;
 
 pub mod error;
+pub mod helpers;
 pub mod instruction;
 pub mod state;
-pub mod helpers;
 
 pinocchio_pubkey::declare_id!("Stake11111111111111111111111111111111111111");
 
-#[cfg(not(feature = "no-entrypoint"))]
-#[panic_handler]
-fn panic(_info: &core::panic::PanicInfo) -> ! {
-    loop {}
-}
+// #[cfg(not(feature = "no-entrypoint"))]
+// #[panic_handler]
+// fn panic(_info: &core::panic::PanicInfo) -> ! {
+//     loop {}
+// }
