+++ program/src/state/mod.rs
@@ -1,6 +1,6 @@
 pub mod accounts;
 
-
+pub mod delegation;
 pub mod stake_flag;
 pub mod stake_history;
 pub mod stake_state_v2;
@@ -8,11 +8,8 @@ pub mod state;
 
 pub use accounts::*;
 
+pub use delegation::*;
 pub use stake_flag::*;
 pub use stake_history::*;
 pub use stake_state_v2::*;
 pub use state::*;
-
-
-
-
