+++ program/src/state/stake_flag.rs
@@ -1,16 +1,11 @@
 #[repr(C)]
 #[derive(Copy, PartialEq, Eq, Clone, PartialOrd, Ord, Hash, Debug)]
 pub struct StakeFlags {
-    bits: u8,
+    pub(crate) bits: u8,
 }
 
 /// Currently, only bit 1 is used. The other 7 bits are reserved for future usage.
 impl StakeFlags {
-    ///  Stake must be fully activated before deactivation is allowed (bit 1).
-    #[deprecated(
-        since = "2.1.0",
-        note = "This flag will be removed because it was only used for `redelegate`, which will not be enabled."
-    )]
     pub const MUST_FULLY_ACTIVATE_BEFORE_DEACTIVATION_IS_PERMITTED: Self =
         Self { bits: 0b0000_0001 };
